requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clients==1.2
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.22.0
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-otp==1.0.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
dynaconf==3.1.0
filelock==3.0.12
gunicorn==20.0.4
h11==0.9.0
httpcore==0.10.2
httpx==0.14.2
hvac==0.10.5
idna==2.10
isort==4.3.21
Jinja2==2.11.2
knox==0.1.0
lazy-object-proxy==1.4.3
loguru==0.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.7
# Pillow==7.2.0
pip-check-reqs==2.1.1
pipenv==2020.8.13
# psycopg2==2.8.5
py4j==0.10.9
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
shortuuid==1.0.1
six==1.15.0
sniffio==1.1.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
stripe==2.50.0
toml==0.10.1
twilio==6.45.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.1.0
win32-setctime==1.0.1
wrapt==1.12.1
